requirements.txt
PyYAML
telethon
cfonts
